contents/serializers.py
from rest_framework import serializers
from contents.models import Content, Media

# Lead Serializer
class ContentSerializer(serializers.ModelSerializer):
    class Meta:
        model = Content
        fields = "__all__"


class MediaSerializer(serializers.ModelSerializer):
    class Meta:
        model = Media
        fields = "__all__"

