requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.1
cryptography==3.4.7
decorator==4.4.2
dj-database-url==0.5.0
Django==3.2.3
django-cors-headers==3.7.0
django-rest-knox==4.1.0
django-s3direct==1.1.8
djangorestframework==3.12.4
dynaconf==3.1.4
gunicorn==20.1.0
hvac==0.10.14
ibm-cloud-sdk-core==3.5.2
ibm-watson==5.1.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.4
django-rest-knox==4.1.0
loguru==0.5.3
moviepy==1.0.3
numpy==1.20.3
Pillow==8.2.0
proglog==0.1.9
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
shortuuid==1.0.1
six==1.16.0
sqlparse==0.4.1
tqdm==4.60.0
urllib3==1.26.4
validators==0.18.2
websocket-client==0.48.0
whitenoise==5.2.0
django-heroku==0.3.1
psycopg2-binary==2.8.6
django-storages==1.7.1
python-dotenv
